+++ Tests/Unit/BackEnd/ModuleTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use TYPO3\CMS\Core\Utility\VersionNumberUtility;
+
 /**
  * Test case.
  *
@@ -14,6 +16,10 @@ class Tx_Seminars_Tests_Unit_BackEnd_ModuleTest extends Tx_Phpunit_TestCase
 
     protected function setUp()
     {
+        if (VersionNumberUtility::convertVersionNumberToInteger(TYPO3_version) >= 8000000) {
+            self::markTestSkipped('This test is for the old BE module only.');
+        }
+
         \Tx_Oelib_ConfigurationProxy::getInstance('seminars')->setAsBoolean('enableConfigCheck', false);
 
         $this->fixture = new Tx_Seminars_BackEnd_Module();